Fused ops need to defuse Move two non-fused ops out of fused
