+++ src/ngraph/runtime/cpu/builder/convolution.cpp
@@ -16,7 +16,7 @@
 
 #include "ngraph/runtime/cpu/kernel/convolution.hpp"
 #include "ngraph/op/fused/conv_fused.hpp"
-#include "ngraph/op/fused/group_conv.hpp"
+#include "ngraph/op/group_conv.hpp"
 #include "ngraph/runtime/cpu/cpu_builder.hpp"
 #include "ngraph/runtime/cpu/mkldnn_invoke.hpp"
 #include "ngraph/runtime/cpu/mkldnn_utils.hpp"
