+++ src/ngraph/runtime/cpu/op/group_conv_bias.hpp
@@ -15,7 +15,7 @@
 //*****************************************************************************
 
 #pragma once
-#include "ngraph/op/fused/group_conv.hpp"
+#include "ngraph/op/group_conv.hpp"
 #include "ngraph/op/op.hpp"
 #include "ngraph/runtime/cpu/cpu_backend_visibility.h"
